fix: resolve AL0086 analyzer errors for GenAI token attributes
Activity.SetTag(string, object?) boxes int arguments, causing AL0086
false positives. Narrow token types from long to int (per OTel semconv),
suppress AL0086 globally in Directory.Build.props, fix source generator
token casts, and fix SpanQueryBuilder lambda type inference.

+++ src/qyl.copilot/Instrumentation/CopilotSpanRecorder.cs
@@ -20,7 +20,7 @@ public static class CopilotSpanRecorder
     /// <param name="activity">The activity to record on.</param>
     /// <param name="inputTokens">Input/prompt tokens.</param>
     /// <param name="outputTokens">Output/completion tokens.</param>
-    public static void RecordTokenUsage(Activity? activity, long inputTokens, long outputTokens)
+    public static void RecordTokenUsage(Activity? activity, int inputTokens, int outputTokens)
     {
         if (activity is null) return;
